Rename GF.Compile.TypeCheck.RConcrete to GF.Compile.TypeCheck.Concrete
